+++ soane/comms/__init__.py
@@ -5,4 +5,5 @@ Package definition for 'soane.comms'.
 from soane.comms._base  import group
 from soane.comms.create import create
 from soane.comms.list   import list_
+from soane.comms.open   import open_
 from soane.comms.read   import read